requirements.txt
openai>=1.35.0
python-dotenv>=1.0.0
requests>=2.31.0
websockets>=12.0
numpy>=1.26.0
sounddevice>=0.4.6